# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Your super  ApplicationAPI",
        "version": "1.0.0"
    },
    "paths": {
...
        "api/searchProizvoditel": {
            "post": {
                "tags": [
                    "Search"
                ],
                "summary": "searchProizvoditel",
                "description": "searchProizvoditel",
                "operationId": "searchProizvoditel",
                "requestBody": {
                    "description": "Pass user credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "NULL"
                                ],
                                "properties": {
                                    "company_name": {
                                        "type": "string",
                                        "format": "name",
                                        "example": "Lait Kuxni"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "api/filterProizvoditel": {
            "post": {
                "tags": [
                    "Search"
                ],
                "summary": "filterProizvoditel",
                "description": "filterProizvoditel",
                "operationId": "filterProizvoditel",
                "requestBody": {
                    "description": "Pass user credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "NULL"
                                ],
                                "properties": {
                                    "meshok": {
                                        "type": "string",
                                        "format": "meshok",
                                        "example": "1 OR 2 OR 3 OR 4"
                                    },
                                    "category_name": {
                                        "type": "string",
                                        "format": "name",
                                        "example": "kuxni"
                                    },
                                    "city_name": {
                                        "type": "string",
                                        "format": "name",
                                        "example": "Moskva^sankt peterburg"
                                    },
                                    "made_in": {
                                        "type": "string",
                                        "format": "name",
                                        "example": "Russia"
                                    },
                                    "show_room": {
                                        "type": "string",
                                        "format": "name",
                                        "example": "Да, Нет"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
...
    }
}